Resolved the conflicts and missing definitions. Happy with the split of app.c
