+++ src/app/app_auth_callbacks.h
@@ -24,9 +24,7 @@
 
 int app_accept_or_reject_session(discovery_service *ds,
     jnx_guid *ig, jnx_guid *sg);
-
 session *app_accept_chat(app_context_t *context);
-
 session *app_reject_chat(app_context_t *context);
 
 #endif // __APP_AUTH_CALLBACKS_H__
